+++ src/layout/routerView/iframes.vue
@@ -1,31 +1,26 @@
 <template>
-	<div>
-		<div class="layout-view-bg-white flex h100" v-loading="iframeLoading">
-			<iframe :src="iframeUrl" frameborder="0" height="100%" width="100%" id="iframe" v-show="!iframeLoading"></iframe>
-		</div>
+	<div class="layout-view-bg-white flex" :style="{ height: `calc(100vh - ${iframeHeight}`, border: 'none' }" v-loading="iframeLoading">
+		<iframe :src="iframeUrl" frameborder="0" height="100%" width="100%" id="iframe" v-show="!iframeLoading"></iframe>
 	</div>
 </template>
 
 <script lang="ts">
-import { defineComponent, reactive, toRefs, onMounted, onBeforeMount, onUnmounted, nextTick, getCurrentInstance } from 'vue';
+import { defineComponent, reactive, toRefs, onMounted, nextTick, watch } from 'vue';
 import { useRoute } from 'vue-router';
+import { useStore } from '/@/store/index';
 export default defineComponent({
 	name: 'layoutIfameView',
-	props: {
-		meta: {
-			type: Object,
-			default: () => {},
-		},
-	},
-	setup(props, { emit }) {
-		const { proxy } = getCurrentInstance() as any;
+	setup() {
 		const route = useRoute();
+		const store = useStore();
 		const state = reactive({
 			iframeLoading: true,
 			iframeUrl: '',
+			iframeHeight: '',
 		});
 		// 初始化页面加载 loading
 		const initIframeLoad = () => {
+			state.iframeUrl = route.meta.isLink;
 			nextTick(() => {
 				state.iframeLoading = true;
 				const iframe = document.getElementById('iframe');
@@ -35,22 +30,24 @@ export default defineComponent({
 				};
 			});
 		};
-		// 页面加载前
-		onBeforeMount(() => {
-			state.iframeUrl = props.meta.isLink;
-			proxy.mittBus.on('onTagsViewRefreshRouterView', (path: string) => {
-				if (route.path !== path) return false;
-				emit('getCurrentRouteMeta');
-			});
-		});
+		// 设置 iframe 的高度
+		const initIframeHeight = () => {
+			let { isTagsview } = store.state.themeConfig.themeConfig;
+			if (isTagsview) return (state.iframeHeight = `84px`);
+			else return (state.iframeHeight = `20px`);
+		};
 		// 页面加载时
 		onMounted(() => {
 			initIframeLoad();
+			initIframeHeight();
 		});
-		// 页面卸载时
-		onUnmounted(() => {
-			proxy.mittBus.off('onTagsViewRefreshRouterView', () => {});
-		});
+		// 监听路由变化，多个 iframe 时使用
+		watch(
+			() => route.path,
+			() => {
+				initIframeLoad();
+			}
+		);
 		return {
 			...toRefs(state),
 		};